Moved from Database to Model package
