+++ app/src/main/java/com/tz/sanga/moviestore/Model/FavoriteDb.java
@@ -1,4 +1,4 @@
-package com.tz.sanga.moviestore.Database;
+package com.tz.sanga.moviestore.Model;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -6,8 +6,6 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.util.Log;
-
-import com.tz.sanga.moviestore.Model.Movie;
 
 import java.util.ArrayList;
 import java.util.List;
